Bringing the school activities to your android phone!
ℹ️ An app especially made to represent the school worldwide!

▶️ See the school activities at your home
You'll get complete details of every event that is happening in the
school through the app interface.Every activity that has happened or is
being live in the school will be visible to all users of the app.User
can view the events details with photos of respective events from the
app and stay updated to the school.

▶️ See the videos of creativities of students
You can view videos of the school and its culture,students work through
this app.You'll receive instant notifications of the new videos that are
being made available through this app everybody.Everyone can see the
videos of creative minds of students as they make something creative and
amazing for all to see!

▶️ Location of the school available to you all the time
Location of the school is available to you, You can view your current
location and the school location for visiting the school.This app makes
it easy to locate the school through the google maps in the app.

🌈 Talent Kids Pre-School
@2018 | School App by VaradParlikar